# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Account": {
				"properties": {
					"homePage": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"homePage",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"LanguageMap": {
				"$ref": "#/components/schemas/Record_string.string_",
				"description": "RFC 5646 language-tagged string map. Keys are language tags, values are human-readable strings."
			},
...
			"InteractionType": {
				"type": "string",
				"enum": [
					"true-false",
					"choice",
					"fill-in",
					"long-fill-in",
					"matching",
					"performance",
					"sequencing",
					"likert",
					"numeric",
					"other"
				]
			},
			"InteractionComponent": {
				"properties": {
					"id": {
						"type": "string"
					},
					"description": {
						"$ref": "#/components/schemas/LanguageMap"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Extensions": {
				"$ref": "#/components/schemas/Record_string.unknown_",
				"description": "Open extension map. Keys are IRIs, values are arbitrary JSON."
			},
			"ActivityDefinition": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/LanguageMap"
					},
					"description": {
						"$ref": "#/components/schemas/LanguageMap"
					},
					"type": {
						"type": "string"
					},
					"moreInfo": {
						"type": "string"
					},
					"interactionType": {
						"$ref": "#/components/schemas/InteractionType"
					},
					"correctResponsesPattern": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"choices": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"scale": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"source": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"target": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/InteractionComponent"
						},
						"type": "array"
					},
					"extensions": {
						"$ref": "#/components/schemas/Extensions"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Activity": {
				"properties": {
					"objectType": {
						"type": "string",
						"enum": [
							"Activity"
						],
						"nullable": false
					},
					"id": {
						"type": "string"
					},
					"definition": {
						"$ref": "#/components/schemas/ActivityDefinition"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Person": {
				"description": "Response from GET /agents — merged agent identity.",
				"properties": {
					"objectType": {
						"type": "string",
						"enum": [
							"Person"
						],
						"nullable": false
					},
					"name": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"mbox": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"mbox_sha1sum": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"openid": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"account": {
						"items": {
							"$ref": "#/components/schemas/Account"
						},
						"type": "array"
					}
				},
				"required": [
					"objectType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"xapi_basic": {
				"type": "http",
				"scheme": "basic"
			}
		}
	},
	"info": {
		"title": "xAPI LRS",
		"version": "2.0.0",
		"contact": {}
	},
	"paths": {
...
		"/xapi/agents": {
			"get": {
				"operationId": "GetAgent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Person"
								}
							}
						}
					}
				},
				"description": "GET /xapi/agents\n\nReturns a Person object — the merged identity information for the\ngiven agent across all Statements in the LRS.",
				"tags": [
					"xAPI Agents"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/xapi/agents/profile": {
			"get": {
				"operationId": "GetAgentProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"xAPI Agent Profile"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "profileId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "since",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PutAgentProfile",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI Agent Profile"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "profileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-None-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostAgentProfile",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI Agent Profile"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "profileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-None-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteAgentProfile",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI Agent Profile"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "profileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/xapi/activities/profile": {
			"get": {
				"operationId": "GetActivityProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"xAPI Activity Profile"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "profileId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "since",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PutActivityProfile",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI Activity Profile"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "profileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-None-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostActivityProfile",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI Activity Profile"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "profileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-None-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteActivityProfile",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"xAPI Activity Profile"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "profileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "If-Match",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/xapi/activities": {
			"get": {
				"operationId": "GetActivity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Activity"
								}
							}
						}
					}
				},
				"description": "GET /xapi/activities\n\nReturns the canonical Activity object for the given activityId.\nPer spec, if the LRS has no canonical definition it returns\nan Activity with just the id.",
				"tags": [
					"xAPI Activities"
				],
				"security": [
					{
						"jwt": []
					},
					{
						"xapi_basic": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "activityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}